Preparation for 0.4.0 Release
