+++ i18n/jpn/package.i18n.json
@@ -7,6 +7,7 @@
     "extension.memoConfig.title": "Memo: 設定",
     "extension.memoReDate.title": "Memo: ファイル名の日付を最新に付け替える",
     "extension.memoTodo.title": "Memo: Todo ",
+    "extension.memoOpenFolder.title": "Memo: メモ格納フォルダを新しい VS Code インスタンスで開く", 
     "memoPath.desc": "memo コマンドへの絶対パス。Serve コマンドを使用する場合に必要",
     "serve-addr.desc": "Memo:Serve コマンドで使用するポート番号を指定する",
     "dateFormat.desc": "date-fns の形式に従ってください。詳細: https://date-fns.org/v1.29.0/docs/format",
