+++ package.nls.json
@@ -7,6 +7,7 @@
     "extension.memoConfig.title": "Memo: Config",    
     "extension.memoReDate.title": "Memo: Re:Date",
     "extension.memoTodo.title": "Memo: Todo",
+    "extension.memoOpenFolder.title": "Memo: Open Memo Folder in New Instance",
     "memoPath.desc": "Absolute path to the memo command",
     "serve-addr.desc": "Specify port number to `Memo: Serve`",
     "dateFormat.desc": "Follow format of date-fns. See: https://date-fns.org/v1.29.0/docs/format",
